% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppfunnel.R
\name{ppfunnel}
\alias{ppfunnel}
\title{Create a new ppfunnel plot}
\usage{
ppfunnel(
  data,
  indiv.data = TRUE,
  outcome = "poisson",
  indicator = "ISR",
  target = 1,
  test = "score",
  method = "FE",
  q.winsor = c(0.05, 0.95),
  levels = c(0.05, 0.01, 0.1),
  flags = c("lower", "expected", "higher"),
  colors = c("blue", "#E69F00", "red"),
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  xexpand = waiver(),
  yexpand = waiver(),
  xbreaks = waiver(),
  ybreaks = waiver(),
  ptsize = 0.8,
  lwd = 0.6,
  title = waiver(),
  ...
)
}
\arguments{
\item{data}{A data frame including summary information used for creating a
graphic. See \code{indiv.data} for details on required columns.}

\item{indiv.data}{logical. If \code{TRUE} (default) then each row of \code{data}
corresponds to an individual (e.g., patient or hospital discharge), and columns
\code{"O"} (observed outcome), \code{"E"} (expected outcome), \code{"ID"} (provider ID) are
required. If \code{FALSE} then each row of \code{data} corresponds to a provider (e.g.,
hospital or facility), and \code{"O"} (observed outcome sum) and \code{"E"} (expected
outcome sum) are required. When \code{outcome = "bernoulli"}, an additional column
\code{"probs"}, each element of which is a list of expected outcomes for a certain
provider, is required.}

\item{outcome}{A character string indicating the outcome type.
If \code{"poisson"} then the outcome is a count (i.e., 0, 1, 2, ...).
If \code{"bernoulli"} then the outcome is a binary indicator (i.e., 0 or 1).
Abbreviation supported.}

\item{indicator}{A character string indicating the indicator of interest.
\code{"ISR"} (indirectly standardized ratio, default) and \code{"Prop"} (proportion) are
supported.}

\item{target}{numeric. It specifies the desired expectation for providers
considered "in-control". If \code{indicator = "ISR"} then \code{target} defaults to 1.}

\item{test}{A character string indicating the type of statistical test used
for calculating the precision and control limits. \code{"score"} (score test,
default) and \code{"exact"} (exact test) are supported.}

\item{method}{A character string indicating the method used for overdispersion
due to incomplete risk adjustment. Options include \code{"FE"} (fixed effects
modeling omitting overdispersion), \code{"MO"} (multiplicative overdispersion without
Winsorization), \code{"MOW"} (multiplicative overdispersion with Winsorization),
\code{"MODW"} (multiplicative overdispersion with debiased Winsorization), \code{"AO"}
(additive overdispersion), \code{"indivEN.exact"} (individualized empirical null
without Taylor approximation), \code{"indivEN.approx"} (individualized empirical
null with Taylor approximation), \code{"indivEN.0meanapprox"} (individualized
empirical null with Taylor approximation and zero mean assumption).}

\item{q.winsor}{A numeric vector specifying the lower and upper thresholds for
Winsorization, with \code{c(0.05, 0.95)} as the default.}

\item{levels}{A numeric vector with 1 to 5 elements specifying the p-value(s)
for displaying control limits, possibly multiple pairs when the length of
\code{levels} is greater than 1.}

\item{flags}{A vector of three character strings for flagging providers, with
\verb{"c("lower", "expected", "higher")"} as the default.}

\item{colors}{A vector of three character strings specifying the colors for
flagging.}

\item{xlab, ylab}{Titles for the \code{x} and \code{y} axes, respectively.}

\item{xlim, ylim}{Numeric vectors of length two providing ranges of the axes
passed on to \code{scale_x_continuous} and \code{scale_y_continuous} of \code{ggplot2}.
\code{NULL} uses the default scale range.}

\item{xexpand, yexpand}{Numeric vectors of length two for adding some padding
to data to ensure that they are placed some distance away from the axes.
\code{xexpand} and \code{yexpand} will be passed on to \code{scale_x_continuous} and
\code{scale_y_continuous} of \code{ggplot2}.}

\item{xbreaks, ybreaks}{Numeric vectors passed on to \code{scale_x_continuous} and
\code{scale_y_continuous} of \code{ggplot2}.}

\item{ptsize}{A positive number specifying the size of the points, passed on
to \code{geom_point} of \code{ggplot2}.}

\item{lwd}{A positive number specifying the line width, passed on to
\code{geom_line} of \code{ggplot2}.}

\item{title}{A character string specifying the text of the title, passed on
to \code{labs} of \code{ggplo2}.}

\item{...}{A list of additional arguments passed on to \code{theme} of \code{ggplot2}.}
}
\value{
A \code{ggplot} object of class \code{"gg"} and \code{"ggplot"}.
}
\description{
\code{ppfunnel()} initializes a ppfunnel plot. It can be used to declare the input
data frame, outcome type, and indicator of interest.
}
\details{
\code{ppfunnel()} is used to create funnel plots for profiling health care
providers. The default \code{method = "FE"} for constructing control limits is
based on a model with fixed provider effects (see
\href{https://doi.org/10.1007/s10985-013-9264-6}{He et al., 2013} for an example
of Bernoulli outcomes). Methods \code{"MO"}, \code{"MOW"}, \code{"MODW"}, \code{"AO"} are based on
\href{https://doi.org/10.1002/sim.1970}{Spiegelhalter (2004)}. Methods based on
the individualized empirical null approach were discussed in
(cite online preprint).
}
\examples{
library(ggplot2)
mytheme <- theme(legend.justification=c(1,1), legend.position=c(1,1),
legend.box="horizontal", legend.text=element_text(size=14),
axis.title=element_text(size=14),
axis.text=element_text(size=14),
plot.title=element_text(hjust=0), text=element_text(size=13))
# Poisson outcome with method = "FE"
SHR \%>\% ppfunnel(indiv.data=F, legend.justification=c(1,1),
legend.position=c(1,1), legend.box="horizontal",
legend.text=element_text(size=14), axis.title=element_text(size=14),
axis.text=element_text(size=14), plot.title=element_text(hjust=0),
text=element_text(size=13))
# alternative use
SHR \%>\% ppfunnel(indiv.data=F) + mytheme
# Poisson outcome with method = "indivEN.0meanapprox"
SHR \%>\% ppfunnel(indiv.data=F, method="indivEN.0meanapprox") + mytheme

}
\references{
Spiegelhalter DJ (2004) "Funnel plots for comparing institutional
performance", \emph{Stat Med} 24(8): 1185--1202 \url{https://doi.org/10.1002/sim.1970}.

He K, Kalbfleisch JD, Li Y, and Li Y (2013) "Evaluating hospital readmission
rates in dialysis facilities; adjusting for hospital effects",
\emph{Lifetime Data Anal} 19(4): 490--512 \url{https://doi.org/10.1007/s10985-013-9264-6}.
}
\author{
Wenbo Wu and Kevin He
}
